training-ground/requirements.txt
kiteconnect==5.0.1
python-dotenv==1.2.1
